requirements.txt
-e .
aws-cdk.aws-eks
pyyaml

